+++ SharedKernel/PuzzleConverter.cs
@@ -148,7 +148,7 @@ namespace AdventOfCode2022.SharedKernel
 
             return adjacent;
         }
-        public static bool isPointInMatrix<T>(T[,] matrix, (int x, int y) point)
+        public static bool isPointInMatrix<T>(T[,] matrix, (Int64 x, Int64 y) point)
         {
             return (point.x >= 0 && point.x < matrix.GetLength(0) && point.y >= 0 && point.y < matrix.GetLength(1));
         }
